[ADD] Configure installation scripts

+++ src/datasets.py
@@ -1,37 +1,53 @@
 import os
+import json
 import shutil
 from roboflow import Roboflow
 
-# API keys e identificadores
 api_key = os.getenv("ROBOFLOW_API_KEY")
 workspace_key = os.getenv("ROBOFLOW_WORKSPACE_KEY")
 project_key = os.getenv("ROBOFLOW_DATASET_KEY")
 
-# Pasta de destino
-output_dir = "data/roboflow"
-os.makedirs(output_dir, exist_ok=True)
+MODO_CLEAN = False
 
-# Inicializa o Roboflow e baixa o dataset
-rf = Roboflow(api_key=api_key)
-dataset = (
-    rf
-    .workspace(workspace_key)
-    .project(project_key)
-    .version(5)  # altere conforme a versão desejada
-    .download("yolov8")
-)
+with open("config.json", "r", encoding="utf-8") as file:
+    CONFIG = json.load(file)
 
-# Caminho do dataset baixado
-download_path = dataset.location
+DATASET_CONFIG = CONFIG["datasets_config"]
 
-# Caminho final desejado
-final_path = os.path.join(output_dir, os.path.basename(download_path))
+names = ["ROBOFLOW_API_KEY", "ROBOFLOW_WORKSPACE_KEY", "ROBOFLOW_DATASET_KEY"]
+values = [api_key, workspace_key, project_key]
+status = True
 
-# Se já existir, remove a antiga antes de mover
-if os.path.exists(final_path):
-    shutil.rmtree(final_path)
+for i, key in enumerate(values):
+    if key == '':
+        print("Erro! chave não encontrada:", names[i])
+        status = False
 
-# Move a pasta baixada para o destino
-shutil.move(download_path, final_path)
+if status:
+    try:
 
-print(f"✅ Dataset movido com sucesso para: {final_path}")
+        output_dir = "data/roboflow" if not MODO_CLEAN else "data/raw"
+        os.makedirs(output_dir, exist_ok=True)
+
+        rf = Roboflow(api_key=api_key)
+        dataset = (
+            rf
+            .workspace(DATASET_CONFIG["roboflow_workspace"])
+            .project(DATASET_CONFIG["roboflow_dataset_name"])
+            .version(DATASET_CONFIG["dataset_version"])  
+            .download(DATASET_CONFIG["download_format_model"])
+        )
+
+        download_path = dataset.location
+
+        final_path = os.path.join(output_dir, os.path.basename(download_path))
+
+        if os.path.exists(final_path):
+            shutil.rmtree(final_path)
+
+        shutil.move(download_path, final_path)
+
+        print(f"✅ Dataset movido com sucesso para: {final_path}")
+
+    except Exception as e:
+        print("Erro ao baixar o dataset:", e)
